Extract sanity check methods to a common class
Since SPI cannot depend on Guava, the sanity checks are usually carried out by
a cloned 'checkArgument' and 'checkState' methods. This commit extracts those
methods into a common helper class that can be used throughout the SPI.
This should slightly decrease the size of the module.
