+++ core/trino-spi/src/main/java/io/trino/spi/Preconditions.java
@@ -1,0 +1,60 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.trino.spi;
+
+import static java.lang.String.format;
+
+/**
+ * This class provides a very limited subset of methods found in Guava's Preconditions class.
+ * The SPI cannot be dependent on Guava since it is too heavy.
+ */
+public final class Preconditions
+{
+    private Preconditions() {}
+
+    public static void checkArgument(boolean expression)
+    {
+        if (!expression) {
+            throw new IllegalArgumentException();
+        }
+    }
+
+    public static void checkArgument(boolean expression, Object errorMessage)
+    {
+        if (!expression) {
+            throw new IllegalArgumentException(String.valueOf(errorMessage));
+        }
+    }
+
+    public static void checkArgument(boolean expression, String errorMessageTemplate, Object... errorMessageArgs)
+    {
+        if (!expression) {
+            throw new IllegalArgumentException(format(errorMessageTemplate, errorMessageArgs));
+        }
+    }
+
+    public static void checkState(boolean expression)
+    {
+        if (!expression) {
+            throw new IllegalStateException();
+        }
+    }
+
+    public static void checkState(boolean expression, String errorMessageTemplate, Object... errorMessageArgs)
+    {
+        if (!expression) {
+            throw new IllegalStateException(format(errorMessageTemplate, errorMessageArgs));
+        }
+    }
+}
