+++ core/trino-spi/src/main/java/io/trino/spi/ptf/ReturnTypeSpecification.java
@@ -13,7 +13,7 @@
  */
 package io.trino.spi.ptf;
 
-import static io.trino.spi.ptf.Preconditions.checkArgument;
+import static io.trino.spi.Preconditions.checkArgument;
 import static java.util.Objects.requireNonNull;
 
 /**
