+++ core/trino-spi/src/main/java/io/trino/spi/type/LongDecimalType.java
@@ -27,6 +27,7 @@ import io.trino.spi.function.ScalarOperator;
 import java.math.BigInteger;
 
 import static io.airlift.slice.SizeOf.SIZE_OF_LONG;
+import static io.trino.spi.Preconditions.checkArgument;
 import static io.trino.spi.function.OperatorType.COMPARISON_UNORDERED_LAST;
 import static io.trino.spi.function.OperatorType.EQUAL;
 import static io.trino.spi.function.OperatorType.XX_HASH_64;
