+++ core/trino-spi/src/main/java/io/trino/spi/type/ShortDecimalType.java
@@ -28,6 +28,7 @@ import java.util.stream.LongStream;
 import java.util.stream.Stream;
 
 import static io.airlift.slice.SizeOf.SIZE_OF_LONG;
+import static io.trino.spi.Preconditions.checkArgument;
 import static io.trino.spi.function.OperatorType.COMPARISON_UNORDERED_LAST;
 import static io.trino.spi.function.OperatorType.EQUAL;
 import static io.trino.spi.function.OperatorType.HASH_CODE;
