+++ core/trino-spi/src/main/java/io/trino/spi/type/TimeZoneKey.java
@@ -31,6 +31,7 @@ import java.util.Properties;
 import java.util.Set;
 import java.util.TreeMap;
 
+import static io.trino.spi.Preconditions.checkArgument;
 import static io.trino.spi.StandardErrorCode.INVALID_FUNCTION_ARGUMENT;
 import static java.lang.Math.abs;
 import static java.lang.Math.max;
@@ -346,12 +347,5 @@ public final class TimeZoneKey
     private static String zoneIdForOffset(long offset)
     {
         return format("%s%02d:%02d", offset < 0 ? "-" : "+", abs(offset / 60), abs(offset % 60));
-    }
-
-    private static void checkArgument(boolean check, String message, Object... args)
-    {
-        if (!check) {
-            throw new IllegalArgumentException(format(message, args));
-        }
     }
 }
